requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
python-multipart
